fix: resolve alliance structure cost overflow for high levels
- Updated AllianceStructureService to calculate costs as float to avoid integer overflow.
- Updated Alliance and AllianceBankLog entities and repositories to support float currency values.
- Implemented DECIMAL(60, 0) in database for bank_credits and log amounts via migration.
- Added string formatting for large numbers in repository SQL bindings to maintain precision.
- Added unit tests to verify high-level structure purchase logic.

+++ app/Models/Entities/Alliance.php
@@ -29,7 +29,7 @@ readonly class Alliance
         public readonly bool $is_joinable,          
         public readonly int $leader_id,
         public readonly int $net_worth,
-        public readonly int $bank_credits,       
+        public readonly float $bank_credits,       
         public readonly ?string $last_compound_at, 
         public readonly string $created_at
     ) {
